Make courses appear in the term details RecyclerView.
SchoolTrackerRepository.java
Remove the mAssociatedCourses declaration from the class. Remove assignment of
mAssociatedCourses in the constructor. Assign this in getmAssociatedCourses().

TermAdapter.java
In onBindViewHolder(), use getTermId() to get the termId and use this in the
intent.putExtra() to pass this to the TermDetails activity.

CourseViewModel.java
Remove mAssociatedCourses assignment from the constructor and move it to
getmAssociatedCourses(). Pass termId to the repository's getAssociatedCourses().

TermDetails.java
Take the termId from the setTermDetails() and pass this to mCourseViewModel. Use
mCourseViewModel to observe to Course data and execute getmAssociatedCourses() with
the termId for setTermDetails().